refactor: reorganize docker structure
