+++ docker/src/assistente2.py
