+++ docker/src/processaimg.py
